backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
alpaca-py==0.33.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
apscheduler==3.10.4
pandas==2.2.3
numpy==2.1.1
praw==7.7.1
tweepy==4.14.0
finnhub-python==2.4.20
requests==2.32.3
textblob==0.18.0
spacy==3.7.6
anthropic>=0.40.0
PyJWT>=2.8.0
cryptography>=42.0.0
